*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@473 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/except.c
@@ -721,7 +721,7 @@ int code;
 {
 	/* Raises an Eiffel exception of the given code with no associated tag */
 
-	eraise((char *) 0, code);
+	eraise((char *) 0, (long) code);
 }
 
 public void eraise(tag, num)
